tout marche

+++ rendu/srcs/requirements/wordpress/tools/start_wp.sh
@@ -2,37 +2,38 @@
 
 sleep 5
 
-#WP_PATH="/var/www/wordpress"
+# WP_PATH="/var/www/wordpress"
 
-# A supprimer
-MYSQL_DATABASE="data_name"
-MYSQL_USER="user_data"
-MYSQL_PASSWORD="mdp"
+echo "Vérification de l'installation de WordPress..."
 
-echo "Installation de WordPress..."
+# Vérifier si le répertoire de WordPress existe
+if [ ! -d "$WP_PATH" ] || [ ! -f "$WP_PATH/wp-config.php" ]; then
+    echo "Installation de WordPress..."
 
-mkdir -p "$WP_PATH"
-    
-chown -R root:root "$WP_PATH"
-wget https://wordpress.org/latest.tar.gz -P /tmp
-tar -xzf /tmp/latest.tar.gz -C /tmp
-mv /tmp/wordpress/* "$WP_PATH"
-rm -rf /tmp/latest.tar.gz /tmp/wordpress
+    mkdir -p "$WP_PATH"
+    chown -R root:root "$WP_PATH"
 
+    # Télécharger et extraire WordPress
+    wget https://wordpress.org/latest.tar.gz -P /tmp
+    tar -xzf /tmp/latest.tar.gz -C /tmp
+    mv /tmp/wordpress/* "$WP_PATH"
+    rm -rf /tmp/latest.tar.gz /tmp/wordpress
 
+    # Créer le fichier wp-config.php si nécessaire
+    if [ ! -f "$WP_PATH/wp-config.php" ]; then
+        wp config create --dbname="$MYSQL_DATABASE" --dbuser="$MYSQL_USER" --dbpass="$MYSQL_PASSWORD" --dbhost="localhost" --path="$WP_PATH"
 
-# # Vérifier si le fichier wp-config.php existe
-# if [ ! -f "$WP_PATH/wp-config.php" ]; then
-#     # Créer le fichier wp-config.php avec WP-CLI
-#     wp config create --dbname="$MYSQL_DATABASE" --dbuser="$MYSQL_USER" --dbpass="$MYSQL_PASSWORD" --dbhost="localhost" --path="$WP_PATH"
-    
-#     # Installer WordPress avec WP-CLI
-#     wp core install --url="http://example.com" --title="Mon Site WordPress" --admin_user="admin" --admin_password="admin_password" --admin_email="email@example.com" --path="$WP_PATH"
+        # Installer WordPress avec WP-CLI
+        wp core install --url="http://example.com" --title="Mon Site WordPress" --admin_user="admin" --admin_password="admin_password" --admin_email="email@example.com" --path="$WP_PATH"
 
-#     echo "Fichier wp-config.php créé et WordPress installé avec succès."
-# else
-#     echo "Le fichier wp-config.php existe déjà."
-# fi
+        echo "Fichier wp-config.php créé et WordPress installé avec succès."
+#        echo "Fichier wp-config.php doit etre créé."
+    else
+        echo "WordPress est déjà installé, fichier wp-config.php trouvé."
+    fi
+else
+    echo "WordPress est déjà installé, répertoire et wp-config.php trouvés."
+fi
 
 # Installation et activation des plugins, si nécessaire
 # wp plugin install wp-redis --activate --path="$WP_PATH"
